JETUT-72: catch up faces hello world as per changes in tutorial; tried mvn install it but GF7 throws error invalid user/password -- WIP
