--- tutorial/web/faces/hello1/src/main/java/jakarta/tutorial/hello1/Hello.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (c), Eclipse Foundation, Inc. and its licensors.
- *
- * All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Distribution License v1.0, which is available at
- * https://www.eclipse.org/org/documents/edl-v10.php
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-package jakarta.tutorial.hello1;
-
-
-import jakarta.enterprise.context.RequestScoped;
-import jakarta.inject.Named;
-
-@Named
-@RequestScoped
-public class Hello {
-
-    private String name;
-
-    public Hello() {
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String user_name) {
-        this.name = user_name;
-    }
-}
-
