--- tutorial/web/faces/hello1/src/main/webapp/index.xhtml
@@ -1,47 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<!--
-
-    Copyright (c), Eclipse Foundation, Inc. and its licensors.
-
-    All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Distribution License v1.0, which is available at
-    https://www.eclipse.org/org/documents/edl-v10.php
-
-    SPDX-License-Identifier: BSD-3-Clause
-
--->
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
-    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html lang="en"
-      xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:h="jakarta.faces.html">
-    <h:head>
-        <title>Facelets Hello Greeting</title>
-    </h:head>
-    <h:body>
-        <h:form>
-            <h:graphicImage url="#{resource['images:duke.waving.gif']}" 
-                            alt="Duke waving his hand"/>
-            <h2>Hello, my name is Duke. What's yours?</h2>
-            <h:inputText id="username"
-                         title="My name is: "
-                         value="#{hello.name}"
-                         required="true"
-                         requiredMessage="Error: A name is required."
-                         maxlength="25" />
-            <p></p>
-            <h:commandButton id="submit" value="Submit" 
-                             action="response" />
-            <h:commandButton id="reset" value="Reset" type="reset" />
-        </h:form>
-        <div class="messagecolor">
-            <h:messages showSummary="true" 
-                        showDetail="false"
-                        errorStyle="color: #d20005" 
-                        infoStyle="color: blue"/>
-        </div>
-    </h:body>
-</html>
-
