--- tutorial/web/faces/hello1/src/main/webapp/response.xhtml
@@ -1,32 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<!--
-
-    Copyright (c), Eclipse Foundation, Inc. and its licensors.
-
-    All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Distribution License v1.0, which is available at
-    https://www.eclipse.org/org/documents/edl-v10.php
-
-    SPDX-License-Identifier: BSD-3-Clause
-
--->
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
-    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html lang="en"
-      xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:h="jakarta.faces.html">
-    <h:head>
-        <title>Facelets Hello Response</title>
-    </h:head>
-    <h:body>
-        <h:form>
-            <h:graphicImage url="#{resource['images:duke.waving.gif']}" 
-                            alt="Duke waving his hand"/>
-            <h2>Hello, #{hello.name}!</h2>
-            <p></p>
-            <h:commandButton id="back" value="Back" action="index" />
-        </h:form>
-    </h:body>
-</html>
