+++ tutorial/web/faces/pom.xml
@@ -32,7 +32,7 @@
     <module>bookmarks</module>
     <module>checkout-module</module>
     <module>guessnumber-faces</module>
-    <module>hello1</module>
+    <module>hello-faces</module>
     <module>hello1-rlc</module>
     <module>reservation</module>
     <module>simple-flow</module>
